Move containers to individual directories
