+++ client/index.dev.jsx
@@ -6,8 +6,8 @@ import store from './src/store/store';
 
 import './src/styles/style.css';
 
-import AvailabilityContainer from './src/containers/Availability';
-import PriceSummary from './src/containers/PriceSummary';
+import AvailabilityContainer from './src/containers/Availability/Availability';
+import PriceSummary from './src/containers/PriceSummary/PriceSummary';
 
 const randomRoom = Math.floor(Math.random() * 100);
