+++ client/src/containers/Availability/Availability.jsx
@@ -2,8 +2,8 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 
-import CalendarsContainer from '../components/Calendars/Calendars';
-import fetchAvailability from '../store/actions/fetchAvailability.action';
+import CalendarsContainer from '../../components/Calendars/Calendars';
+import fetchAvailability from '../../store/actions/fetchAvailability.action';
 
 export class Availability extends Component {
   componentDidMount() {
