+++ client/src/containers/PriceSummary/PriceSummary.spec.jsx
